+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Notes
 
 - Testing automations
+- Corrected Constructor standings tab after Favorite Drivers implementation broke it
 
 ---